pyproject.toml
[project]
name = "autotaskllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "db-sqlite3>=0.0.1",
    "duckdb>=1.2.0",
    "faker>=36.1.0",
    "fastapi>=0.115.8",
    "gitpython>=3.1.44",
    "markdown>=3.7",
    "numpy>=2.2.2",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sqlite-utils>=3.38",
    "uvicorn>=0.34.0",
    "whisper>=1.1.10",
]
